refactor(hooks): add usePortal other folder
